Finished webpack config
