--- app/assets/stylesheets/search.css
@@ -1,136 +1,0 @@
-.nav-search {
-  -webkit-flex-grow: 1;
-  -ms-flex-positive: 1;
-  position: relative;
-  width: 60vw;
-  max-width: 80vw
-  height: 47px;
-}
-
-.nav-search.active {
-  box-shadow: 0 4px 4px rgba(79, 79, 79, 0.21);
-}
-.nav-search.active .search-dropdown {
-  display: block;
-}
-
-.nav-search.active .search-input {
-  -webkit-animation: expand-search-box-animation 0.5s forwards;
-  animation: expand-search-box-animation 0.5s forwards;
-}
-
-.nav-search.active .search-input input {
-  border-width: 2px;
-}
-
-.nav-search.active .search-input .close-search {
-  display: inline-block;
-}
-
-.nav-search.active .search-input .search-dropdown {
-  display: block;
-}
-
-.nav-search .search-input {
-  transition: left 0.2s ease-in-out;
-  transition: width 0s ease-in-out;
-}
-
-.nav-search .search-input .search-icon {
-  position: absolute;
-  left: 15px;
-  top: 13px;
-  z-index: 0;
-  color: #4FBADC;
-}
-
-.nav-search .search-input input {
-  height: 50px;
-  border: 1px solid #84D9F5;
-  border-radius: 4px;
-  min-width: 300px;
-  width: 100%;
-  padding-left: 50px;
-}
-
-.nav-search .search-input input:focus {
-  outline: none;
-}
-
-.nav-search .search-input i.close-search {
-  color: #84D9F5;
-  display: none;
-  position: absolute;
-  right: 15px;
-  top: 13px;
-  cursor: pointer;
-}
-
-.search-dropdown {
-  box-sizing: border-box;
-  color: #B3B3B3;
-  opacity: 1.00;
-  padding: 20px;
-  width: 100%;
-  -webkit-animation: expand-search-dropdown-animation 0.5s forwards;
-  animation: expand-search-dropdown-animation 0.5s forwards;
-  overflow-y: scroll;
-  max-height: 400px;
-  border-radius: 0 0 4px 4px;
-  background-color: #FCFCFC;
-  border: 1px solid #E0E0E0;
-  box-shadow: 1px 3px 4px rgba(0, 0, 0, 0.09);
-  display: none;
-}
-
-.search-dropdown .small {
-  -webkit-flex-basis: 35%;
-  -ms-flex-preferred-size: 35%;
-  flex-basis: 35%;
-}
-
-.search-dropdown .search-section .hits-blank {
-  color: #666;
-  text-align: center;
-  padding-top: 20px;
-}
-
-.search-dropdown a {
-  text-decoration: none;
-  color: inherit;
-}
-
-.hit {
-  border-bottom: 1px solid #E6E6E6;
-  margin-bottom: 20px;
-}
-.hit .hit-title {
-  color: #44A948;
-  border-bottom: 1px solid #44A948;
-  font-weight: 500;
-  margin-bottom: 0;
-  margin-top: 0;
-  display: inline-block;
-  font-size: 14px;
-}
-.hit .hit-description {
-  text-decoration: none;
-  color: black;
-  font-size: 14px;
-  display: block;
-  margin-top: 3px;
-}
-.hit .hit-anchor {
-  font-size: 13px;
-  color: #666;
-}
-.hit .algolia-docsearch-suggestion--highlight {
-  background-color: #FFE9A4;
-}
-.hit:last-child {
-  /*border: 0;*/
-}
-
-.ais-hits--item:last-child .hit {
-  border: 0;
-}
