--- app/dashboard/dashboard.css
@@ -1,3 +1,0 @@
-@import url(../assets/stylesheets/application.css);
-@import url(../assets/stylesheets/footer.css);
-@import url(../assets/stylesheets/sidebar.css);
