--- app/dashboard/dashboard.html
@@ -1,9 +1,0 @@
-<html>
-<head>
-  <meta charset="UTF-8">
-  <title>Document</title>
-</head>
-<body>
-
-</body>
-</html>
