--- app/home-page/home.css
@@ -1,30 +1,0 @@
-@import url(../assets/stylesheets/search.css);
-@import url(../assets/stylesheets/application.css);
-@import url(../assets/stylesheets/footer.css);
-@import url(../assets/stylesheets/sidebar.css);
-
-.home-banner {
-  justify-content: center;
-  align-items: center;
-  display: flex;
-}
-
-.home-h1 {
-  color: #56ccf2;
-  font-weight: lighter;
-  letter-spacing: 5px;
-  margin-top: 20vh;
-  font-size: 30px;
-  margin-bottom: 20px;
-  text-align: center;
-}
-
-body {
-  font-family: "SourceHan";
-  background-color: rgb(250, 250, 250);
-}
-
-.container {
-  min-height: 100vh;
-}
-
