+++ dashboard.css
@@ -1,0 +1,3 @@
+@import url(assets/stylesheets/application.css);
+@import url(assets/stylesheets/footer.css);
+@import url(assets/stylesheets/sidebar.css);
