+++ index.html
@@ -1,0 +1,73 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <title>DianDianTou</title>
+
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/all.css" integrity="sha384-hWVjflwFxL6sNzntih27bfxkr27PmbbK/iSvJ+a4+0owXq79v+lsFkW54bOGbiDQ" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+    <link rel="stylesheet" type="text/css" href="index.css">
+
+
+  </head>
+  <body>
+    <div id="menuToggle">
+    <!--
+    A fake / hidden checkbox is used as click reciever,
+    so you can use the :checked selector on it.
+    -->
+      <input type="checkbox" />
+
+    <!--
+    Some spans to act as a hamburger.
+
+    They are acting like a real hamburger,
+    not that McDonalds stuff.
+    -->
+      <span></span>
+      <span></span>
+      <span></span>
+
+    <!--
+    Too bad the menu has to be inside of the button
+    but hey, it's pure CSS magic.
+    -->
+      <ul id="menu">
+        <a href="#"><li><i class="fas fa-search"></i>  搜索</li></a>
+        <a href="#"><li><i class="fas fa-chart-pie"></i>   面板</li></a>
+      </ul>
+    </div>
+    <div class="container">
+      <div class= "home-banner">
+        <div class="home-banner-content">
+          <h1 class="home-h1">投资从这里开始...</h1>
+          <div class="nav-search">
+            <div class="search-input">
+              <i class="material-icons search-icon">search</i>
+              <input type="text" placeholder="请输入公司名称" id="input-content" />
+              <div class="search-dropdown" id="hits-container">
+                <ul id="results">
+  <!--       <li>Resultat 1 (TODO: remove)</li>
+        <li>Resultat 2 (TODO: remove)</li> -->
+                </ul>
+              </div>
+              <i class="close-search material-icons">close</i>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="footer-container">
+      <div class="footer">
+        <div class="footer-links">
+          <a href="#"><p class="foot-text">关于我们</p></a>
+          <a href="#"><p class="foot-text">联系我们</p></a>
+        </div>
+      </div>
+    </div>
+
+    <script src="build/application.js"></script>
+  </body>
+</html>
