+++ package.json
@@ -1,0 +1,20 @@
+{
+  "name": "javascript-101",
+  "version": "1.0.0",
+  "description": "Starting code for First JavaScript ES6 lecture",
+  "main": "index.js",
+  "repository": "https://github.com/lewagon/javascript-101",
+  "author": "ssaunier",
+  "license": "MIT",
+  "dependencies": {
+    "axios": "^0.18.0",
+    "webpack": "^4.8.1",
+    "webpack-cli": "^2.1.3",
+    "webpack-dev-server": "^3.1.4"
+  },
+  "devDependencies": {
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-plugin-import": "^2.11.0"
+  }
+}
