+++ src/browser.js
@@ -1,0 +1,5 @@
+import axios from 'axios'
+import './search.js'
+import './application.js'
+import './cable.js'
+
