+++ webpack.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  entry: "./src/browser.js",
+  mode: "development",
+  output: {
+    filename: "build/application.js"
+  },
+  devtool: "sourcemap"
+};
